inference/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-inference"
version = "0.1.0"
description = "AI Inference Microservice for asynchronous image processing with face detection and embedding generation"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "pydantic>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.24.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "sse-starlette>=1.6.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[tool.setuptools.packages.find]
include = ["src*"]
exclude = ["tests*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "*/conftest.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
